added the Developer.md file and refactored test files
