+++ src/test/java/com/laserfiche/api/client/unit/OAuthUtilTest.java
@@ -1,4 +1,4 @@
-package com.laserfiche.api.client.test;
+package com.laserfiche.api.client.unit;
 
 import com.laserfiche.api.client.integration.BaseTest;
 import org.junit.jupiter.api.Test;
